+++ src/Tabby/Core/Application.cpp
@@ -28,6 +28,8 @@ Application::Application(const ApplicationSpecification& specification)
     m_Window = Window::Create(WindowProps(m_Specification.Name, m_Specification.Width, m_Specification.Height, m_Specification.MinWidth, m_Specification.MinHeight, m_Specification.FullscreenMode, m_Specification.Resizable, m_Specification.VSync));
     m_Window->SetEventCallback(TB_BIND_EVENT_FN(Application::OnEvent));
 
+    Renderer::Init();
+
     m_ImGuiRenderer = new ImGuiRenderer;
     m_ImGuiRenderer->Launch(m_Window->GetNativeWindow());
 }
@@ -38,6 +40,7 @@ Application::~Application()
 
     m_LayerStack.Destroy();
 
+    Renderer::Shutdown();
     // AudioEngine::Shutdown();
     AssetManager::FullUnload();
 }